another round of refactoring
